fix: api auth no cache
