+++ frontend-client/lib/utils.ts
@@ -1,5 +1,7 @@
+import { authOptions } from '@/app/api/auth/[...nextauth]/route'
 import { type ClassValue, clsx } from 'clsx'
 import ky from 'ky'
+import { getServerSession } from 'next-auth'
 import { twMerge } from 'tailwind-merge'
 import { baseUrl } from './vars'
 
@@ -10,4 +12,17 @@ export const cn = (...inputs: ClassValue[]) => {
 export const fetcher = ky.create({
   prefixUrl: baseUrl,
   throwHttpErrors: false
+})
+
+export const fetcherWithAuth = fetcher.extend({
+  hooks: {
+    beforeRequest: [
+      async (request) => {
+        const session = await getServerSession(authOptions)
+        if (session) {
+          request.headers.set('Authorization', session.token.accessToken)
+        }
+      }
+    ]
+  }
 })
