+++ frontend-client/next.config.js
@@ -3,7 +3,15 @@ const nextConfig = {
   experimental: {
     typedRoutes: true
   },
-  output: 'standalone'
+  output: 'standalone',
+  async headers() {
+    return [
+      {
+        source: '/api/auth/:path*',
+        headers: [{ key: 'Cache-Control', value: 'no-store, max-age=0' }]
+      }
+    ]
+  }
 }
 
 module.exports = nextConfig
